Synchronize plugin: Do not turn an SSH path into an absolute path
SSH paths of the form user@host:/remote/path can be used as either src or dest with rsync.

As the synchronize module already uses the presence of ':' to detect a source/destination over SSH, this commit updates the action plugin accordingly.
Without this commit, the path is turned into /local/path/user@host:/remote/path, which is not a valid syntax for rsync.